대쉬히트 하면 마나회복하는거 구현 완료

+++ API_FrameWork/CplayerData.h
@@ -7,15 +7,7 @@ class CplayerData: public Singleton<CplayerData>
 {
 	//공격력 마법력 공속 경험치
 	//스킬쿨감수치
-	//이속 최대체력 최대마나 
-	
-	//장비의 마나는 마지막에 사용됨
-	//장비마나까지 소비하면 빈마나상태가됨
-	//빈마나인 장비를 마나가 잇을때 착용시 장비의 빈마나를 현재마나로 채움
-	//마나의소비는 오른쪽에서 왼쪽으로
-
-	//빈하트의 장비를 체력이 잇을때 착용시 체력을 뺏어감
-	//소비는 오른쪽에서 왼쪽
+	//이속
 
 private:
 	bool _isDebug;
@@ -44,7 +36,7 @@ private:
 	int _presentMP;
 	int _MaxMP;
 
-	int _defaultStamina;
+	float _defaultStamina;
 	int _recoveryStaminaCoolTimeCount;
 	
 	int _EXP;
